Basic GraphQl products page
